Use of size_t causes an ominous warning: specified bound between 18446744071562067968 and 18446744073709551615 exceeds maximum object size 9223372036854775807 [-Wstringop-overflow=]
in GPU builds with CUDA 12. Changing to int since these are small loops, which removes the warning.

+++ zero/gkyl_dg_diffusion_gyrokinetic_priv.h
@@ -12,9 +12,9 @@ static inline int diffdirs_linidx(const bool *isdirdiff, int cdim) {
   // kernel for each combination of diffusive directions).
   bool diff_in_dir[GKYL_MAX_CDIM];
   if (isdirdiff)
-    for (size_t d=0; d<cdim; d++) diff_in_dir[d] = isdirdiff[d];
+    for (int d=0; d<cdim; d++) diff_in_dir[d] = isdirdiff[d];
   else
-    for (size_t d=0; d<cdim; d++) diff_in_dir[d] = true;
+    for (int d=0; d<cdim; d++) diff_in_dir[d] = true;
 
   // Linear index into list of volume kernels.
   int dirs_bin_key[] = {1,2,4,8,16,32}; // Binary: 000001, 000010, 000100, 001000, 010000, 100000.
